Moved inline container drop handling to InlineContainerView. Added a dropzone for when the container is empty.
Added styles for the dropzone.
