+++ src/sheets/container/ContainerContentsSections.svelte
@@ -41,8 +41,6 @@
     TidyFlags.sectionConfig.get(container)?.[CONSTANTS.TAB_CONTAINER_CONTENTS],
   );
 
-  $: inlineContainerToggleServiceStore = inlineContainerToggleService.store;
-
   let itemIdsToShow =
     getContext<Readable<Set<string> | undefined>>('itemIdsToShow');
 
@@ -73,25 +71,13 @@
   $: gridTemplateColumns = `/* Name */ 1fr /* Weight */ 3rem /* Quantity */ 3rem ${classicControlsWidth}`;
 
   const localize = FoundryAdapter.localize;
-
-  async function onDrop(
-    event: DragEvent & { currentTarget: EventTarget & HTMLElement },
-  ) {
-    container.sheet._onDrop(event);
-
-    event.preventDefault();
-    event.stopImmediatePropagation();
-  }
 </script>
 
 {#each configuredContents as section (section.key)}
   {#if section.show}
-    <!-- TODO: Apply proper a11y trappings for this -->
-    <!-- svelte-ignore a11y-no-static-element-interactions -->
     <section
       class="container-contents-list-section"
       style="--grid-template-columns: {gridTemplateColumns};"
-      on:drop={onDrop}
     >
       <TidyTable key={section.key}>
         <svelte:fragment slot="header">
@@ -134,7 +120,10 @@
                   {item}
                 />
                 {#if 'containerContents' in ctx && !!ctx.containerContents}
-                  <InlineContainerToggle {item} {inlineContainerToggleService} />
+                  <InlineContainerToggle
+                    {item}
+                    {inlineContainerToggleService}
+                  />
                 {/if}
                 <!-- This is generally what we want in Tidy Tables / Item Table V2; consider breaking off ItemNameV2 to propagate and replace the old ItemName gradually. -->
                 <ItemName
