+++ src/sheets/container/InlineContainerView.svelte
@@ -6,6 +6,7 @@
   import { getContext } from 'svelte';
   import type { Readable } from 'svelte/store';
   import type { ContainerContents, Item5e } from 'src/types/item.types';
+  import { FoundryAdapter } from 'src/foundry/foundry-adapter';
 
   export let container: Item5e;
   export let containerContents: ContainerContents;
@@ -18,15 +19,27 @@
 
   let itemIdsToShow =
     getContext<Readable<Set<string> | undefined>>('itemIdsToShow');
+
+  async function onDrop(
+    event: DragEvent & { currentTarget: EventTarget & HTMLElement },
+  ) {
+    container.sheet._onDrop(event);
+
+    event.preventDefault();
+    event.stopImmediatePropagation();
+  }
 </script>
 
 <ExpandableContainer
   expanded={$inlineContainerToggleServiceStore.has(container.id)}
   class={!!$itemIdsToShow && !$itemIdsToShow.has(container.id) ? 'hidden' : ''}
 >
+  <!-- TODO: Apply proper a11y trappings for this -->
+  <!-- svelte-ignore a11y-no-static-element-interactions -->
   <div
     class="flex-column extra-small-gap flex-1 inline-container-view"
     data-tidy-container-id={container.id}
+    on:drop={onDrop}
   >
     <CapacityBar {container} capacity={containerContents.capacity} />
     <ContainerContentsSections
@@ -38,5 +51,12 @@
       {inlineContainerToggleService}
       {sheetDocument}
     />
+    {#if !containerContents.contents.some((c) => c.items.length > 0)}
+      <div class="empty-container">
+        <span class="empty-container-text"
+          >{FoundryAdapter.localize('TIDY5E.EmptyContainer')}</span
+        >
+      </div>
+    {/if}
   </div>
 </ExpandableContainer>
